Update test mocks
